[Runtimes] Move runtime handlers to the API
