+++ mlrun/api/runtime_handlers/remotesparkjob.py
@@ -1,0 +1,29 @@
+# Copyright 2023 Iguazio
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+from mlrun.api.runtime_handlers.kubejob import KubeRuntimeHandler
+from mlrun.runtimes.base import RuntimeClassMode
+
+
+class RemoteSparkRuntimeHandler(KubeRuntimeHandler):
+    kind = "remote-spark"
+    class_modes = {RuntimeClassMode.run: "remote-spark"}
+
+    @staticmethod
+    def _are_resources_coupled_to_run_object() -> bool:
+        return True
+
+    @staticmethod
+    def _get_object_label_selector(object_id: str) -> str:
+        return f"mlrun/uid={object_id}"
