+++ tests/api/runtime_handlers/test_daskjob.py
@@ -17,8 +17,9 @@ from kubernetes import client
 from sqlalchemy.orm import Session
 
 import mlrun.common.schemas
+from mlrun.api.runtime_handlers import get_runtime_handler
 from mlrun.api.utils.singletons.db import get_db
-from mlrun.runtimes import RuntimeKinds, get_runtime_handler
+from mlrun.runtimes import RuntimeKinds
 from mlrun.runtimes.constants import PodPhases
 from tests.api.runtime_handlers.base import TestRuntimeHandlerBase
